// src/types.ts
export interface PurchaseState {
  sourceData: Array<any>;
  selectedSource: any | null;
  grnData: Array<any>;
  selectedGrn: any | null;
  itemData: Array<any>;
  selectedItem: any | null;
  itemDetails: Array<any>;
  additionalCostData: Array<any>;
  selectedAdditionalCost: any | null;
  capitalisationData: string | null;
  putToUse: string | null;
  manual: boolean;
}

export type PurchaseActionType = {
  type:
    | "sourceData"
    | "selectedSource"
    | "grnData"
    | "selectedGrn"
    | "itemData"
    | "selectedItem"
    | "itemDetails"
    | "additionalCostData"
    | "selectedAdditionalCost"
    | "capitalisationData"
    | "manual"
    | "putToUse";
  payload: any;
};

export interface PurchaseDetails {
  budgetType: string;
  oldUnitId: string;
  vendorCode: string;
  vendorName: string;
  poDate: any;
  poNo: string;
  poSno: string;
  itemCode: string;
  itemName: string;
  grnNo: string;
  grnSno: string;
  grnDate: string;
  qcCompleted: string;
  acceptedQty: number;
  purchaseValue: number;
  grnValue: number;
  billNo: string;
  billDate: any;
  uom: string;
  binLocation: string;
  pjYear: string;
  pjDocId: string;
  pjDocSr: string;
  pjDocNo: string;
}

export interface AdditionalCost {
  id: number;
  costType: any;
  amount: string;
  journalNo: string;
}